refactor: split program EIR lowering
