+++ src/ir_lower/program/spl_discovery.rs
@@ -1,0 +1,352 @@
+//! Purpose:
+//! Builtin SPL method reachability and dynamic-construction metadata.
+//!
+//! Called from:
+//! - `crate::ir_lower::program`.
+//!
+//! Key details:
+//! - Keeps program metadata deterministic and EIR lowering behavior unchanged.
+
+use super::*;
+
+/// Lowers the small builtin SPL method slice currently consumed by the EIR backend.
+pub(super) fn lower_referenced_builtin_spl_methods(
+    module: &mut Module,
+    check_result: &CheckResult,
+    constants: &std::collections::HashMap<String, (ExprKind, PhpType)>,
+    fiber_return_sigs: &std::collections::HashMap<String, crate::types::FunctionSig>,
+) {
+    loop {
+        let mut methods = referenced_builtin_spl_methods(module);
+        methods.sort();
+        methods.dedup();
+        methods.retain(|(class_name, method_key)| {
+            !class_method_already_lowered(module, class_name, method_key, false)
+                && !runtime_intrinsic_method_has_wrapper(class_name, method_key, false)
+        });
+        if methods.is_empty() {
+            break;
+        }
+
+        let before = module.class_methods.len();
+        for (class_name, method_key) in methods {
+            lower_builtin_spl_method(
+                &class_name,
+                &method_key,
+                module,
+                check_result,
+                constants,
+                fiber_return_sigs,
+            );
+        }
+        for method in module.class_methods.iter_mut().skip(before) {
+            method.flags.is_synthetic = true;
+        }
+        if module.class_methods.len() == before {
+            break;
+        }
+    }
+}
+
+/// Finds builtin SPL methods whose symbols are required by already-lowered EIR.
+pub(super) fn referenced_builtin_spl_methods(module: &Module) -> Vec<(String, String)> {
+    let mut methods = Vec::new();
+    for function in module
+        .functions
+        .iter()
+        .chain(module.class_methods.iter())
+        .chain(module.closures.iter())
+        .chain(module.fiber_wrappers.iter())
+        .chain(module.callback_wrappers.iter())
+        .chain(module.extern_callback_trampolines.iter())
+        .chain(module.runtime_callable_invokers.iter())
+    {
+        for inst in &function.instructions {
+            match inst.op {
+                Op::ObjectNew => {
+                    if let Some(class_name) = class_data_name(module, inst) {
+                        let construct_key = php_method_key("__construct");
+                        push_supported_builtin_spl_method_for_receiver(
+                            &mut methods,
+                            module,
+                            class_name,
+                            &construct_key,
+                        );
+                        push_builtin_spl_metadata_methods(&mut methods, module, class_name);
+                    }
+                }
+                Op::DynamicObjectNew => {
+                    if let Some((fallback_class, required_parent)) =
+                        dynamic_object_new_metadata_names(module, inst)
+                    {
+                        let construct_key = php_method_key("__construct");
+                        if is_supported_builtin_spl_method(fallback_class, &construct_key) {
+                            methods.push((fallback_class.to_string(), construct_key.clone()));
+                        }
+                        if is_supported_builtin_spl_method(required_parent, &construct_key) {
+                            methods.push((required_parent.to_string(), construct_key));
+                        }
+                        push_builtin_spl_metadata_methods(&mut methods, module, fallback_class);
+                        push_builtin_spl_metadata_methods(&mut methods, module, required_parent);
+                    }
+                }
+                Op::DynamicObjectNewMixed => {
+                    let construct_key = php_method_key("__construct");
+                    for class_name in module.class_infos.keys() {
+                        if !is_dynamic_new_mixed_metadata_candidate(class_name) {
+                            continue;
+                        }
+                        push_supported_builtin_spl_method_for_receiver(
+                            &mut methods,
+                            module,
+                            class_name,
+                            &construct_key,
+                        );
+                        push_builtin_spl_metadata_methods(&mut methods, module, class_name);
+                    }
+                }
+                Op::DynamicObjectNewWithoutConstructorMixed => {
+                    for class_name in module.class_infos.keys() {
+                        if !is_dynamic_new_mixed_metadata_candidate(class_name) {
+                            continue;
+                        }
+                        push_builtin_spl_metadata_methods(&mut methods, module, class_name);
+                    }
+                }
+                Op::MethodCall | Op::NullsafeMethodCall => {
+                    let Some(receiver) = inst.operands.first().copied() else {
+                        continue;
+                    };
+                    let Some(receiver_ty) = function
+                        .value(receiver)
+                        .map(|value| value.php_type.codegen_repr())
+                    else {
+                        continue;
+                    };
+                    let Some(method_name) = string_data_name(module, inst) else {
+                        continue;
+                    };
+                    let method_key = php_method_key(method_name);
+                    match receiver_ty {
+                        PhpType::Object(class_name) => {
+                            let normalized = class_name.trim_start_matches('\\');
+                            push_supported_builtin_spl_method_for_receiver(
+                                &mut methods,
+                                module,
+                                normalized,
+                                &method_key,
+                            );
+                        }
+                        // A Mixed/Union receiver dispatches at runtime over every class whose
+                        // flattened method set contains this name (mirrors `mixed_method_candidates`
+                        // in the EIR backend). Register the builtin SPL implementation behind each
+                        // candidate so its vtable slot is emitted; otherwise the runtime class-id
+                        // dispatch jumps through a null vtable slot and segfaults. This covers
+                        // method calls on a `mixed` value and on foreach values from object
+                        // iterators (e.g. DirectoryIterator), which the EIR lowers as Mixed locals.
+                        PhpType::Mixed | PhpType::Union(_) => {
+                            for (candidate_class, class_info) in &module.class_infos {
+                                if class_info.methods.contains_key(&method_key) {
+                                    push_supported_builtin_spl_method_for_receiver(
+                                        &mut methods,
+                                        module,
+                                        candidate_class,
+                                        &method_key,
+                                    );
+                                }
+                            }
+                        }
+                        _ => {}
+                    }
+                }
+                _ => {}
+            }
+        }
+    }
+    methods
+}
+
+/// Returns true when generic `new $class` can emit static metadata for this class.
+pub(super) fn is_dynamic_new_mixed_metadata_candidate(class_name: &str) -> bool {
+    if class_name.starts_with("__Elephc") {
+        return false;
+    }
+    if supported_dynamic_new_builtin_class_name(class_name) {
+        return true;
+    }
+    !known_dynamic_new_builtin_class_name(class_name)
+}
+
+/// Returns true for builtin classes with safe static allocation paths in generic dynamic new.
+pub(super) fn supported_dynamic_new_builtin_class_name(class_name: &str) -> bool {
+    matches!(
+        php_symbol_key(class_name.trim_start_matches('\\')).as_str(),
+        "arrayiterator"
+            | "arrayobject"
+            | "badfunctioncallexception"
+            | "badmethodcallexception"
+            | "callbackfilteriterator"
+            | "domainexception"
+            | "error"
+            | "exception"
+            | "fiber"
+            | "fibererror"
+            | "invalidargumentexception"
+            | "iteratoriterator"
+            | "jsonexception"
+            | "lengthexception"
+            | "logicexception"
+            | "outofboundsexception"
+            | "outofrangeexception"
+            | "overflowexception"
+            | "rangeexception"
+            | "recursivecallbackfilteriterator"
+            | "runtimeexception"
+            | "spldoublylinkedlist"
+            | "splfixedarray"
+            | "splqueue"
+            | "splstack"
+            | "typeerror"
+            | "underflowexception"
+            | "unexpectedvalueexception"
+            | "valueerror"
+            | "stdclass"
+    )
+}
+
+/// Returns true for builtin classes that generic dynamic new must not treat as user classes.
+pub(super) fn known_dynamic_new_builtin_class_name(class_name: &str) -> bool {
+    matches!(
+        php_symbol_key(class_name.trim_start_matches('\\')).as_str(),
+        "appenditerator"
+            | "arrayiterator"
+            | "arrayobject"
+            | "badfunctioncallexception"
+            | "badmethodcallexception"
+            | "cachingiterator"
+            | "callbackfilteriterator"
+            | "directoryiterator"
+            | "domainexception"
+            | "emptyiterator"
+            | "error"
+            | "exception"
+            | "fiber"
+            | "fibererror"
+            | "filesystemiterator"
+            | "filteriterator"
+            | "generator"
+            | "globiterator"
+            | "infiniteiterator"
+            | "internaliterator"
+            | "invalidargumentexception"
+            | "iteratoriterator"
+            | "jsonexception"
+            | "lengthexception"
+            | "limititerator"
+            | "logicexception"
+            | "multipleiterator"
+            | "norewinditerator"
+            | "outofboundsexception"
+            | "outofrangeexception"
+            | "overflowexception"
+            | "parentiterator"
+            | "phar"
+            | "phardata"
+            | "rangeexception"
+            | "recursivearrayiterator"
+            | "recursivecachingiterator"
+            | "recursivecallbackfilteriterator"
+            | "recursivedirectoryiterator"
+            | "recursivefilteriterator"
+            | "recursiveiteratoriterator"
+            | "recursiveregexiterator"
+            | "reflectionattribute"
+            | "reflectionclass"
+            | "reflectionmethod"
+            | "reflectionparameter"
+            | "reflectionproperty"
+            | "regexiterator"
+            | "runtimeexception"
+            | "spldoublylinkedlist"
+            | "splfileinfo"
+            | "splfileobject"
+            | "splfixedarray"
+            | "splheap"
+            | "splmaxheap"
+            | "splminheap"
+            | "splobjectstorage"
+            | "splpriorityqueue"
+            | "splqueue"
+            | "splstack"
+            | "spltempfileobject"
+            | "typeerror"
+            | "underflowexception"
+            | "unexpectedvalueexception"
+            | "valueerror"
+            | "stdclass"
+    )
+}
+
+/// Adds the supported builtin SPL method owner for a receiver class or one of its parents.
+pub(super) fn push_supported_builtin_spl_method_for_receiver(
+    methods: &mut Vec<(String, String)>,
+    module: &Module,
+    class_name: &str,
+    method_key: &str,
+) {
+    let mut current = Some(class_name);
+    while let Some(name) = current {
+        if is_supported_builtin_spl_method(name, method_key) {
+            methods.push((name.to_string(), method_key.to_string()));
+            return;
+        }
+        current = module
+            .class_infos
+            .get(name)
+            .and_then(|class_info| class_info.parent.as_deref());
+    }
+}
+
+/// Returns the class-name immediate attached to an instruction.
+pub(in crate::ir_lower) fn class_data_name<'a>(
+    module: &'a Module,
+    inst: &crate::ir::Instruction,
+) -> Option<&'a str> {
+    let Some(Immediate::Data(data)) = inst.immediate else {
+        return None;
+    };
+    module
+        .data
+        .class_names
+        .get(data.as_raw() as usize)
+        .map(String::as_str)
+}
+
+/// Parses dynamic object factory fallback and required-parent metadata.
+pub(in crate::ir_lower) fn dynamic_object_new_metadata_names<'a>(
+    module: &'a Module,
+    inst: &crate::ir::Instruction,
+) -> Option<(&'a str, &'a str)> {
+    class_data_name(module, inst)?.split_once('|')
+}
+
+/// Returns the string immediate attached to an instruction.
+pub(in crate::ir_lower) fn string_data_name<'a>(
+    module: &'a Module,
+    inst: &crate::ir::Instruction,
+) -> Option<&'a str> {
+    let Some(Immediate::Data(data)) = inst.immediate else {
+        return None;
+    };
+    module
+        .data
+        .strings
+        .get(data.as_raw() as usize)
+        .map(String::as_str)
+}
+
+/// Normalizes a PHP method name for metadata lookups.
+pub(in crate::ir_lower) fn php_method_key(method_name: &str) -> String {
+    crate::names::php_symbol_key(method_name)
+}
+
